+++ plugin/includes/init.php
@@ -4,4 +4,5 @@
 $plugin_dir = plugin_dir_path( __FILE__ );
 
 require $plugin_dir . 'utils.php';
+require $plugin_dir . 'curl.php';
 require $plugin_dir . 'zaprite_api.php';
